+++ index.html
@@ -2,6 +2,7 @@
 <html>
     <head lang="en">
         <title>Fixmaster</title>
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link
             rel="stylesheet"
             href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
