+++ src/pages/SalonPage.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import { Link, Params, useParams } from 'react-router-dom';
 import SaloonCard from '../Components/SalonCard/SalonCard';
-import salonsData from '../salonsData';
-import { Accordion, Button, Container } from 'react-bootstrap';
-import ServiceItem, { MyContainer } from '../Components/AccordeonItems/ServiceItem';
-import MasterItem from '../Components/AccordeonItems/MasterItem';
+import { Button, Container } from 'react-bootstrap';
 import styled from 'styled-components';
+import SalonPageAccordion from '../Components/AccordeonItems/SalonPageAccordion';
 
 const Center = styled.div`
     width: 100%;
@@ -22,34 +20,9 @@ const SalonPage = (): JSX.Element => {
                 <h2>Salon Page</h2>
                 <Button variant='outline-dark'><Link to='/'>Главная </Link></Button>
                 <SaloonCard id={id}/>
-                <MyContainer>
-                    <Accordion defaultActiveKey='0'>
-                        <Accordion.Item eventKey='0'>
-                            <Accordion.Header>Мастера салона</Accordion.Header>
-                            <Accordion.Body>
-                                {salonsData[Number( id ) - 1].masters.map( el => {
-									return <div>
-                                        <MasterItem image={el.image} name={el.name}/>
-                                    </div>;
-								} )}
-                            </Accordion.Body>
-                        </Accordion.Item>
-                    </Accordion>
-                </MyContainer>
-                <MyContainer>
-                    <Accordion>
-                        <Accordion.Item eventKey='0'>
-                            <Accordion.Header>Услуги салона</Accordion.Header>
-                            <Accordion.Body>
-                                {salonsData[Number( id ) - 1].services.map( el => {
-									return <div>
-                                        <ServiceItem text={el.title} price={el.price} description={el.description} minTime={el.min_time}/>
-                                    </div>;
-								} )}
-                            </Accordion.Body>
-                        </Accordion.Item>
-                    </Accordion>
-                </MyContainer>
+
+                <SalonPageAccordion/>
+
                 <Center>
                     <Link to={`/record/${id}`}>
                         <Button variant='info'>Записаться онлайн</Button>
